Moved code to new library: ResultStatisticsCache
Moved  RigStatisticsCalculator.h/.cpp, RigStatisticsDataCache.h/.cpp,
RigStatisticsMath.h/.cpp to new library ResultStatisticsCache.
ResInsight and some unit tests now link with this new library.
