+++ ApplicationCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -24,7 +24,6 @@ ${CEE_CURRENT_LIST_DIR}RigMainGrid.h
 ${CEE_CURRENT_LIST_DIR}RigReservoirBuilderMock.h
 ${CEE_CURRENT_LIST_DIR}RigCaseCellResultsData.h
 ${CEE_CURRENT_LIST_DIR}RigSingleWellResultsData.h
-${CEE_CURRENT_LIST_DIR}RigStatisticsMath.h
 ${CEE_CURRENT_LIST_DIR}RigWellPath.h
 ${CEE_CURRENT_LIST_DIR}RigFault.h
 ${CEE_CURRENT_LIST_DIR}RigNNCData.h
@@ -33,8 +32,6 @@ ${CEE_CURRENT_LIST_DIR}cvfGeometryTools.inl
 ${CEE_CURRENT_LIST_DIR}RigPipeInCellEvaluator.h
 ${CEE_CURRENT_LIST_DIR}RigResultAccessor2d.h
 ${CEE_CURRENT_LIST_DIR}RigTernaryResultAccessor2d.h
-${CEE_CURRENT_LIST_DIR}RigStatisticsDataCache.h
-${CEE_CURRENT_LIST_DIR}RigStatisticsCalculator.h
 ${CEE_CURRENT_LIST_DIR}RigNativeStatCalc.h
 ${CEE_CURRENT_LIST_DIR}RigMultipleDatasetStatCalc.h
 )
@@ -58,14 +55,11 @@ ${CEE_CURRENT_LIST_DIR}RigMainGrid.cpp
 ${CEE_CURRENT_LIST_DIR}RigReservoirBuilderMock.cpp
 ${CEE_CURRENT_LIST_DIR}RigCaseCellResultsData.cpp
 ${CEE_CURRENT_LIST_DIR}RigSingleWellResultsData.cpp
-${CEE_CURRENT_LIST_DIR}RigStatisticsMath.cpp
 ${CEE_CURRENT_LIST_DIR}RigWellPath.cpp
 ${CEE_CURRENT_LIST_DIR}RigFault.cpp
 ${CEE_CURRENT_LIST_DIR}RigNNCData.cpp
 ${CEE_CURRENT_LIST_DIR}cvfGeometryTools.cpp
 ${CEE_CURRENT_LIST_DIR}RigTernaryResultAccessor2d.cpp
-${CEE_CURRENT_LIST_DIR}RigStatisticsDataCache.cpp
-${CEE_CURRENT_LIST_DIR}RigStatisticsCalculator.cpp
 ${CEE_CURRENT_LIST_DIR}RigNativeStatCalc.cpp
 ${CEE_CURRENT_LIST_DIR}RigMultipleDatasetStatCalc.cpp
 )
