+++ ApplicationCode/ResultStatisticsCache/RigStatisticsCalculator.cpp
@@ -1,0 +1,44 @@
+/////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (C) Statoil ASA
+//  Copyright (C) Ceetron Solutions AS
+// 
+//  ResInsight is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+// 
+//  ResInsight is distributed in the hope that it will be useful, but WITHOUT ANY
+//  WARRANTY; without even the implied warranty of MERCHANTABILITY or
+//  FITNESS FOR A PARTICULAR PURPOSE.
+// 
+//  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html> 
+//  for more details.
+//
+/////////////////////////////////////////////////////////////////////////////////
+
+#include "RigStatisticsCalculator.h"
+
+#include <cmath> // Needed for HUGE_VAL on Linux
+
+
+//--------------------------------------------------------------------------------------------------
+/// 
+//--------------------------------------------------------------------------------------------------
+void RigStatisticsCalculator::meanCellScalarValue(double& meanValue)
+{
+    double valueSum = 0.0;
+    size_t sampleCount = 0;
+
+    this->valueSumAndSampleCount(valueSum, sampleCount);
+
+    if (sampleCount == 0)
+    {
+        meanValue = HUGE_VAL;
+    }
+    else
+    {
+        meanValue = valueSum / sampleCount;
+    }
+}
+
