+++ ApplicationCode/ResultStatisticsCache/RigStatisticsCalculator.h
