+++ ApplicationCode/ResultStatisticsCache/RigStatisticsDataCache.cpp
