+++ ApplicationCode/ResultStatisticsCache/RigStatisticsMath.cpp
