+++ ApplicationCode/ResultStatisticsCache/RigStatisticsMath.h
